project/requirements.txt
fastapi
uvicorn
python-multipart
openai
pillow
python-dotenv